requirements.txt
python-dotenv == 1.0.1
azure-ai-documentintelligence == 1.0.0b4
injector == 0.22.0
PyMuPDF == 1.25.1
langchain == 0.3.12
langchain-community == 0.3.12
langchain-openai == 0.2.12
